Merge Kubeadm config file with flags into a single config file.
As of Kubeadm 1.8.x, if both `--config` and `--apiserver-cert-extra-sans`
flags are passed to kubeadm init/json, the config file will take precedence.
This is not enough because we need to merge sans detected from node into
the config. These new commands will override non-zero values in --config file
with values passed as flag and print the merged config to StdOut.